refactor: changed some names
